+++ app/page.tsx
@@ -1,5 +1,7 @@
+import BackToTop from "@/components/back-to-top";
 import Footer from "@/components/footer";
 import Navbar from "@/components/navbar";
+import ScrollProgress from "@/components/scroll-progress";
 import About from "@/components/sections/about";
 import Contact from "@/components/sections/contact";
 import Experience from "@/components/sections/experience";
@@ -10,6 +12,7 @@ import Skills from "@/components/sections/skills";
 export default function Home() {
   return (
     <>
+      <ScrollProgress />
       <Navbar />
       <main>
         <Hero />
@@ -20,6 +23,7 @@ export default function Home() {
         <Contact />
       </main>
       <Footer />
+      <BackToTop />
     </>
   );
 }